AddPolygon() Fixed And Major Fixes To Collision
In AddPolygon() the wrong Lua script was being run, so it was creating a
sprite, that was fixed. All objects were not being tested against
each-other for collisions, this was patched with the addition of some
methods to Scan_Area that allow you to find what object(s) collided with
the current object. This required the addition of the IDXYZ class.