Huge Code Clean - Up
Most of the code was cleaned up and now conforms to one style. This
badly needed to be done, a lot of the code was done when I was first
learning C++ and did not have a solid style yet, not to mention my style
"changed" about three times through - out the project, and Scan Areas
were developed independently so they had a completely different "style."
Speaking of which, most of the Scan Area code has finally been removed
and is deprecated until further notice. I also fixed a bad memory leak
that I discovered in components, cleaned up a lot of the "while" loops
that should have been for loops, and the reference to the engine inside
of entities is no longer in an std::vector. Over - all the code is now
somewhat acceptable, and readable. This is something I wanted to do
before White - Storm: Lightning (alpha) was released, and am happy to
release it in a legible state.